rearranging functions and adding a test
