+++ scripts/tests/test_convert_surface.py
@@ -24,3 +24,15 @@ def test_execution_surface_vtk_fib(script_runner):
     ret = script_runner.run('scil_convert_surface.py', in_surf,
                             'rhpialt.ply')
     assert ret.success
+
+
+def test_execution_surface_vtk_xfrom(script_runner):
+    os.chdir(os.path.expanduser(tmp_dir.name))
+    in_surf = os.path.join(get_home(), 'surface_vtk_fib',
+                           'lh.pialt_xform')
+    x_form = os.path.join(get_home(), 'surface_vtk_fib',
+                          'log.txt')
+    ret = script_runner.run('scil_convert_surface.py', in_surf,
+                            'lh.pialt_xform.vtk', '--xform', x_form,
+                            '--to_lps')
+    assert ret.success
